feat(clean): 替换清理脚本，增强功能以支持删除指定文件并处理 mix 目录 #65
