+++ package.json
@@ -9,7 +9,8 @@
 		"build": "node ./scripts/merge-styles.js --dir src/styles && tsc -noEmit -skipLibCheck && node esbuild.config.mjs production",
 		"version": "node version-bump.mjs && git add manifest.json versions.json",
 		"mix": "node ./scripts/mix-code.js --dir ../src",
-		"clean": "node ./scripts/clean-mix.js",
+		"clean": "node ./scripts/clean.js",
+		"clean-build": "node ./scripts/clean.js --delete styles.css data.json main.js",
 		"test": "vitest",
 		"test:watch": "vitest --watch",
 		"test:coverage": "vitest --coverage",
